Use shared_ptr's for passing events instead of references
This makes it easier on the users to use the dispatcher module so they don't
have to maintain the lifecycle of events. Without this, users wouldn't be able
to create events on the fly (e.g. as stack variables), but with this, users can
create events in any context without having to worry about the event lifecycle.